requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-cncf-kubernetes==8.2.0
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.3.4
psycopg[binary]==3.2.1
pytest==8.3.2
